+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Default of 0 waiting time in tick settings.
 - Move functions of layer swapping instead of moving.
 - Crash when syncing a label and removing it afterwards.
+- Cursor visible function just being the cursor grab function.
 
 ## [0.10.0] - 2024-2-10